pyproject.toml
[tool.poetry]
name = "lessmore"
version = "0.3.0"
description = "Mark Lidenberg's Monorepo"
authors = [ "marklidenberg <marklidenberg@gmail.com>" ]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"

[tool.black]
line-length = 120

[tool.isort]
# isort configuration, compatible with black
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.semantic_release]
version_toml = "pyproject.toml:tool.poetry.version"
build_command = false
branch = "master"
